requirements.txt
rembg==2.0.50
Pillow==10.0.0
requests==2.31.0
